Creazione di una immagine per l'esecuzione dei batch calcolo statistiche

+++ build_image.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -x
 
 function printHelp() {
 echo "Usage $(basename $0) [ -t <repository>:<tagname> | <Installer Sorgente> | <Personalizzazioni> | <Avanzate> | -h ]"
@@ -9,13 +9,13 @@ echo "Options
 -h             : Mostra questa pagina di aiuto
 
 Installer Sorgente:
--v <VERSIONE>  : Imposta la versione dell'installer binario da utilizzare per il build (default: ${LATEST_GOVPAY_RELEASE})
+-v <VERSIONE>  : Imposta la versione dell'installer binario da utilizzare per il build (default: ${LATEST_GOVWAY_RELEASE})
 -l <FILE>      : Usa un'installer binario sul filesystem locale (incompatibile con -j)
 -j             : Usa l'installer prodotto dalla pipeline jenkins https://jenkins.link.it/govway-testsuite/installer/govway-installer-<version>.tgz
 
 Personalizzazioni:
 -d <TIPO>      : Prepara l'immagine per essere utilizzata su un particolare database  (valori: [ hsql, postgresql, oracle] , default: hsql)
--a <TIPO>      : Imposta quali archivi inserire nell'immmagine finale (valori: [runtime , manager, all] , default: all)
+-a <TIPO>      : Imposta quali archivi inserire nell'immmagine finale (valori: [runtime , manager, batch, all] , default: all)
 -e <PATH>      : Imposta il path interno utilizzato per i file di configurazione di govway 
 -f <PATH>      : Imposta il path interno utilizzato per i log di govway
 -s             : Prepara anche l'immagine con il software di generazione statistiche
@@ -49,9 +49,9 @@ CUSTOM_MANAGER=
 CUSTOM_WIDLFLY_CLI=
 
 LATEST_LINK="$(curl -qw '%{redirect_url}\n' https://github.com/link-it/govway/releases/latest 2> /dev/null)"
-LATEST_GOVPAY_RELEASE="${LATEST_LINK##*/}"
+LATEST_GOVWAY_RELEASE="${LATEST_LINK##*/}"
 
-while getopts "ht:v:d:jl:i:a:r:m:w:o:e:f:s" opt; do
+while getopts "ht:v:d:jl:i:a:r:m:w:o:e:f:" opt; do
   case $opt in
     t) TAG="$OPTARG"; NO_COLON=${TAG//:/}
       [ ${#TAG} -eq ${#NO_COLON} -o "${TAG:0:1}" == ':' -o "${TAG:(-1):1}" == ':' ] && { echo "Il tag fornito \"$TAG\" non utilizza la sintassi <repository>:<tagname>"; exit 2; } ;;
@@ -66,7 +66,7 @@ while getopts "ht:v:d:jl:i:a:r:m:w:o:e:f:s" opt; do
     i) TEMPLATE="${OPTARG}"
         [ ! -f "${TEMPLATE}" ] && { echo "Il file indicato non esiste o non e' raggiungibile [${TEMPLATE}]."; exit 3; } 
         ;;
-    a) ARCHIVI="${OPTARG}"; case "$ARCHIVI" in runtime);;manager);;all);;*) echo "Tipologia archivi da inserire non riconosciuta: ${ARCHIVI}"; exit 2;; esac ;;
+    a) ARCHIVI="${OPTARG}"; case "$ARCHIVI" in runtime);;manager);;batch);;all);;*) echo "Tipologia archivi da inserire non riconosciuta: ${ARCHIVI}"; exit 2;; esac ;;
     r) CUSTOM_RUNTIME="${OPTARG}"
         [ ! -d "${CUSTOM_RUNTIME}" ] && { echo "la directory indicata non esiste o non e' raggiungibile [${CUSTOM_RUNTIME}]."; exit 3; }
         [ -z "$(ls -A ${CUSTOM_RUNTIME})" ] && { echo "la directory [${CUSTOM_RUNTIME}] e' vuota.";  }
@@ -85,7 +85,6 @@ while getopts "ht:v:d:jl:i:a:r:m:w:o:e:f:s" opt; do
         ;;
     e) CUSTOM_GOVWAY_HOME="${OPTARG}" ;;
     f) CUSTOM_GOVWAY_LOG="${OPTARG}" ;;
-    s) STATS_MODE="batch";;
     h) printHelp
        exit 0
        ;;
@@ -95,19 +94,18 @@ while getopts "ht:v:d:jl:i:a:r:m:w:o:e:f:s" opt; do
       ;;
   esac
 done
-
+[ "${ARCHIVI}" == 'batch' -a "${DB:-hsql}" == 'hsql' ] && { echo "Il build dell'immagine batch non puo' essere eseguita per il database HSQL"; exit 4; }
 
 rm -rf buildcontext
 mkdir -p buildcontext/
 cp -fr commons buildcontext/
 
 DOCKERBUILD_OPT=()
-DOCKERBUILD_OPTS=(${DOCKERBUILD_OPTS[@]} '--build-arg' "govway_fullversion=${VER:-${LATEST_GOVPAY_RELEASE}}")
+DOCKERBUILD_OPTS=(${DOCKERBUILD_OPTS[@]} '--build-arg' "govway_fullversion=${VER:-${LATEST_GOVWAY_RELEASE}}")
 [ -n "${DB}" ] && DOCKERBUILD_OPTS=(${DOCKERBUILD_OPTS[@]} '--build-arg' "govway_database_vendor=${DB}")
 [ -n "${TEMPLATE}" ] &&  cp -f "${TEMPLATE}" buildcontext/commons/
 [ -n "${CUSTOM_GOVWAY_HOME}" ] && DOCKERBUILD_OPTS=(${DOCKERBUILD_OPTS[@]} '--build-arg' "govway_home=${CUSTOM_GOVWAY_HOME}")
 [ -n "${CUSTOM_GOVWAY_LOG}" ] && DOCKERBUILD_OPTS=(${DOCKERBUILD_OPTS[@]} '--build-arg' "govway_log=${CUSTOM_GOVWAY_LOG}")
-[ -n "${STATS_MODE}" ] && DOCKERBUILD_OPTS=(${DOCKERBUILD_OPTS[@]} '--build-arg' "govway_stats_mode=${STATS_MODE}")
 if [ -n "${CUSTOM_RUNTIME}" ]
 then
   cp -r ${CUSTOM_RUNTIME}/ buildcontext/runtime
@@ -133,19 +131,19 @@ fi
 
 
 "${DOCKERBIN}" build "${DOCKERBUILD_OPTS[@]}" \
-  -t linkitaly/govway-installer_${DB:-hsql}:${VER:-${LATEST_GOVPAY_RELEASE}} \
+  -t linkitaly/govway-installer_${DB:-hsql}:${VER:-${LATEST_GOVWAY_RELEASE}} \
   -f ${INSTALLER_DOCKERFILE} buildcontext
 RET=$?
 [ ${RET} -eq  0 ] || exit ${RET}
  
-# Build imagine govway
+# Build imagini GovWAY
 
 [ -n "${ARCHIVI}" ] && DOCKERBUILD_OPTS=(${DOCKERBUILD_OPTS[@]} '--build-arg' "govway_archives_type=${ARCHIVI}")
 if [ -z "$TAG" ] 
 then
   REPO=linkitaly/govway
-  TAGNAME=${VER:-${LATEST_GOVPAY_RELEASE}}
-  [ -n "${ARCHIVI}" -a "${ARCHIVI}" != 'all' ] && TAGNAME=${VER:-${LATEST_GOVPAY_RELEASE}}_${ARCHIVI}
+  TAGNAME=${VER:-${LATEST_GOVWAY_RELEASE}}
+  [ -n "${ARCHIVI}" -a "${ARCHIVI}" != 'all' ] && TAGNAME=${VER:-${LATEST_GOVWAY_RELEASE}}_${ARCHIVI}
   
   # mantengo i nomi dei tag compatibili con quelli usati in precedenza
   case "${DB:-hsql}" in
@@ -167,6 +165,20 @@ then
   DOCKERBUILD_OPTS=(${DOCKERBUILD_OPTS[@]} '--build-arg' "oracle_custom_jdbc=custom_oracle_jdbc")
 fi
 
+
+
+if [ "${ARCHIVI}" == 'batch' ]
+then
+
+  "${DOCKERBIN}" build "${DOCKERBUILD_OPTS[@]}" \
+  --build-arg source_image=linkitaly/govway-installer_${DB:-hsql} \
+  -t "${TAG}" \
+  -f govway/Dockerfile.govway_batch buildcontext
+  RET=$?
+  [ ${RET} -eq  0 ] || exit ${RET}
+
+else
+
 "${DOCKERBIN}" build "${DOCKERBUILD_OPTS[@]}" \
   --build-arg source_image=linkitaly/govway-installer_${DB:-hsql} \
   -t "${TAG}" \
@@ -175,18 +187,12 @@ RET=$?
 [ ${RET} -eq  0 ] || exit ${RET}
 
 
-
-if [ "${STATS_MODE}" == 'batch' ]
-then
-
-  "${DOCKERBIN}" build "${DOCKERBUILD_OPTS[@]}" \
-  --build-arg source_image=linkitaly/govway-installer_${DB:-hsql} \
-  -t "${TAG}_batch" \
-  -f govway/Dockerfile.govway_batch buildcontext
-  RET=$?
-  [ ${RET} -eq  0 ] || exit ${RET}
 fi
-if [ "${DB:-hsql}" != 'hsql' ]
+
+
+
+
+if [ "${DB:-hsql}" != 'hsql' -a "${ARCHIVI}" != 'batch' ]
 then
   mkdir -p compose/govway_{conf,log}
   chmod 777 compose/govway_{conf,log}
@@ -215,6 +221,20 @@ EOYAML
         - GOVWAY_DB_USER=govway
         - GOVWAY_DB_PASSWORD=govway
         - GOVWAY_POP_DB_SKIP=false
+# Decommentare dopo il build dell'immagine batch (usando l'opzione "-a batch")
+#  batch_stat_orarie:
+#    container_name: govway_batch_${SHORT}
+#    image: linkitaly/govway:${VER:-${LATEST_GOVWAY_RELEASE}}_batch_postgres
+#    #command: Giornaliere
+#    #command: Orarie # << default
+#    depends_on:
+#        - database
+#    environment:
+#        - GOVWAY_STAT_DB_SERVER=pg_govway_${SHORT}
+#        - GOVWAY_STAT_DB_NAME=govwaydb
+#        - GOVWAY_STAT_DB_USER=govway
+#        - GOVWAY_STAT_DB_PASSWORD=govway
+#        - GOVWAY_BATCH_USA_CRON=yes
   database:
     container_name: pg_govway_${SHORT}
     image: postgres:13
@@ -253,8 +273,27 @@ EOSQL
         # il container oracle puo impiegare anche 20 minuti ad avviarsi
         - GOVWAY_LIVE_DB_CHECK_MAX_RETRY=120
         - GOVWAY_READY_DB_CHECK_MAX_RETRY=600
+# Decommentare dopo il build dell'immagine batch (usando l'opzione "-a batch")
+#  batch_stat_orarie:
+#    container_name: govway_batch_${SHORT}
+#    image: linkitaly/govway:${VER:-${LATEST_GOVWAY_RELEASE}}_batch_oracle
+#    #command: Giornaliere
+#    #command: Orarie # << default
+#    depends_on:
+#        - database
+#    volumes:
+#        # Il driver deve essere compiato manualmente nella directory corrente
+#        - ./ojdbc10.jar:/tmp/ojdbc10.jar
+#    environment:
+#        - GOVWAY_STAT_DB_SERVER=or_govway_${SHORT}
+#        - GOVWAY_STAT_DB_NAME=GOVWAYPDB
+#        - GOVWAY_STAT_DB_USER=GOVWAY
+#        - GOVWAY_STAT_DB_PASSWORD=GOVWAY
+#        - GOVWAY_ORACLE_JDBC_PATH=/tmp/ojdbc10.jar
+#        - GOVWAY_ORACLE_JDBC_URL_TYPE=servicename
+#        - GOVWAY_BATCH_USA_CRON=yes
   database:
-    container_name: or_govway_${LATEST_GOVPAY_RELEASE}_oracle
+    container_name: or_govway_${SHORT}
     image: container-registry.oracle.com/database/enterprise:19.3.0.0
     shm_size: 2g
     environment:
